requirements.txt
asgiref==3.7.2
Django==4.2.7
django-guardian==2.4.0
sqlparse==0.4.4
